refactor(fee-payer): extract shared test helpers
Move tempoChain, userAccount, sponsorAddress, testMnemonic,
tempoTransport, feePayerTransport, and buildSponsorClient into
test/helpers.ts. The two test files no longer duplicate this setup.
