+++ apps/fee-payer/test/e2e.test.ts
@@ -2,29 +2,15 @@ import { env, exports } from 'cloudflare:workers'
 import { Mnemonic } from 'ox'
 import { createClient, custom, http, parseUnits } from 'viem'
 import { sendTransactionSync } from 'viem/actions'
-import { tempo, tempoDevnet, tempoLocalnet, tempoModerato } from 'viem/chains'
 import { Account, Actions, withRelay } from 'viem/tempo'
 import { beforeAll, describe, expect, it } from 'vitest'
-
-const tempoChain = (() => {
-	const tempoEnv = env.TEMPO_ENV ?? 'localnet'
-	if (tempoEnv === 'moderato' || tempoEnv === 'testnet') return tempoModerato
-	if (tempoEnv === 'mainnet') return tempo
-	if (tempoEnv === 'devnet') return tempoDevnet
-	return tempoLocalnet
-})()
-
-const testMnemonic =
-	'test test test test test test test test test test test junk'
-
-const sponsorAddress = '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266'
-
-const userAccount = Account.fromSecp256k1(
-	Mnemonic.toPrivateKey(testMnemonic, {
-		as: 'Hex',
-		path: Mnemonic.path({ account: 9 }),
-	}),
-)
+import {
+	sponsorAddress,
+	tempoChain,
+	tempoTransport,
+	testMnemonic,
+	userAccount,
+} from './helpers.js'
 
 function createFeePayerTransportWithSpy() {
 	const requests: Array<{ method: string; params: unknown }> = []
@@ -57,31 +43,6 @@ function createFeePayerTransportWithSpy() {
 	})
 
 	return { transport, requests }
-}
-
-function createTempoTransport() {
-	return custom({
-		async request({ method, params }) {
-			const response = await fetch(env.TEMPO_RPC_URL, {
-				method: 'POST',
-				headers: { 'Content-Type': 'application/json' },
-				body: JSON.stringify({
-					jsonrpc: '2.0',
-					id: 1,
-					method,
-					params,
-				}),
-			})
-			const data = (await response.json()) as {
-				result?: unknown
-				error?: { code: number; message: string }
-			}
-			if (data.error) {
-				throw new Error(data.error.message || 'RPC Error')
-			}
-			return data.result
-		},
-	})
 }
 
 // Mint liquidity for fee tokens.
@@ -190,7 +151,7 @@ describe('fee-payer integration', () => {
 			const client = createClient({
 				account: userAccount,
 				chain: tempoChain,
-				transport: withRelay(createTempoTransport(), feePayerTransport, {
+				transport: withRelay(tempoTransport(), feePayerTransport, {
 					policy: 'sign-only',
 				}),
 			})
@@ -228,7 +189,7 @@ describe('fee-payer integration', () => {
 			const client = createClient({
 				account: userAccount,
 				chain: tempoChain,
-				transport: withRelay(createTempoTransport(), feePayerTransport, {
+				transport: withRelay(tempoTransport(), feePayerTransport, {
 					policy: 'sign-and-broadcast',
 				}),
 			})
